Separate StartupModal into two tabs
