+++ package.json
@@ -38,6 +38,7 @@
     "rc-slider": "^8.6.13",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
+    "react-tabs": "^3.1.0",
     "react-test-renderer": "^16.11.0",
     "react-use-websocket": "^0.9.8"
   },
