+++ src/components/StartupModal/TableSelect.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { css } from "@emotion/core";
+
+const TableSelect = () => {
+  return (
+    <div>
+      <h2>Select a Table to Join</h2>
+    </div>
+  );
+};
+
+export default TableSelect;
